update hàm random dề

+++ app/Http/Controllers/QuizController.php
@@ -134,13 +134,23 @@ class QuizController extends Controller
 
     protected function configuration(Request $request)
     {
+
         // $objective_id = $request->objective_id;
 
         // $student_session_id = $request->student_session_id;
 
         $objective_id = 2;
-        $student_session_id = 305;
+        $student_session_id = 291284;
 
+        //hàm dùng chung
+        function select(array $array, $column)
+        {
+            $a = [];
+            foreach ($array as $ss) {
+                array_push($a, $ss[$column]);
+            }
+            return $a;
+        }
 
         // Tìm ra cấu hình
         $config = QuizConfigObjective::where('objective_id', $objective_id)
@@ -152,8 +162,10 @@ class QuizController extends Controller
             array_push($list_domain, $d->subject);
         }
         $domain =  array_flip(array_flip($list_domain));
+        // dd($domain);
 
         $questions = array_flip($list_domain);
+        // dd($questions);
         $tp_qt = TopicQuestion::query()
             ->join('lms_questions', 'lms_topic_question.question_id', 'lms_questions.id');
         // dd($toan);
@@ -171,10 +183,27 @@ class QuizController extends Controller
             'quizz_code' => random_int(1000, 9999),
             'quiz_config_id' => $config->id,
             'student_session_id' => $student_session_id,
-            'available_date' => $student_session->from
+            // 'available_date' => $student_session->from
         ];
 
         $quiz =   Quiz::create($quizzes);
+        //tìm ra học sinh và các câu hỏi đã thi
+        {
+            $student_id = StudentSession::find($student_session_id)->student_id;
+            $student_session = StudentSession::where('student_id', $student_id)->get()->toArray();
+            $arr_student_session_id = select($student_session, 'id');
+
+            $attempts = Attempt::whereIn('student_session', $arr_student_session_id)->get()->toArray();
+            $arr_attempt_id = select($attempts, 'id');
+            // các câu hỏi mà học sinh đó đã thi
+            $attempt_detail = AttemptDetail::whereIn('attempt_id', $arr_attempt_id)
+                ->join('lms_questions', 'lms_attempt_details.question_id', 'lms_questions.id')->distinct()->get();
+
+            $attempt_question = select($attempt_detail->toArray(), 'question_id');
+            // $a = select($attempt_detail->whereNotNull('ref_question_id')->toArray(), 'ref_question_id');
+            // dd($a);
+            // dd(array_unique($a));
+        }
         foreach ($domain as $do) {
             // dd($do);
             $tp_cf = '';
@@ -191,14 +220,19 @@ class QuizController extends Controller
                     if ($qt->question_type == 'complex') {
                         $q =  TopicQuestion::where('topic_id', $qt->topic_id)
                             ->join('lms_question_objective', 'lms_topic_question.question_id', 'lms_question_objective.question_id')
-
                             ->join('lms_questions', 'lms_topic_question.question_id', 'lms_questions.id')
                             ->where('complex', 'sub')->where('objective_id', $objective_id)
                             ->get();
                         $r =  array_unique(array_column($q->toArray(), 'ref_question_id'));
                         array_push($rqi, $r);
+
                         if ($main_id == null || in_array($main_id, $rqi) == false) {
-                            $rand = array_rand($q->toArray(), 1);
+
+                            //Mảng chưa id các câu hỏi main  có câu sub thi
+                            $attempt_detail_main = $attempt_detail->whereNotNull('ref_question_id')->toArray();
+                            $main_exam_id = array_unique(select($attempt_detail_main, 'ref_question_id'));
+
+                            $rand = array_rand($q->whereNotIn('ref_question_id', $main_exam_id)->toArray(), 1);
                             $main_id = $q[$rand]->ref_question_id;
                             // echo "test";
                             // echo $main_id, '<br/>';
@@ -224,7 +258,6 @@ class QuizController extends Controller
                             array_push($arr_q, $q[$rand]);
                         } else {
 
-                            // $q_s = $q->where('ref_question_id', $main_id)->whereNotIn('question_id', $sub_id);
                             $q_s = Question::where('ref_question_id', $main_id)->whereNotIn('id', $sub_id)->get();
                             $rand = array_rand($q_s->toArray(), 1);
                             array_push($sub_id, $q_s[$rand]->id);
@@ -254,26 +287,29 @@ class QuizController extends Controller
                             ->join('lms_questions', 'lms_topic_question.question_id', 'lms_questions.id')
                             ->whereNull('complex')->where('objective_id', $objective_id)
                             ->get();
-                        // dd($q);
-                        // echo "<pre>";
-                        // print_r($q->toArray());
-                        $rand = array_rand($q->toArray(), 1);
+                        $q_id = select($q->toArray(), 'id');
+
+
+                        // cát lấy mảng các câu hỏi chưa thi để random
+                        $at_n_questtion = array_diff($q_id, $attempt_question);
+                        $rand = array_rand($at_n_questtion, 1);
                         $q_q = [
-                            'question_id' => $q[$rand]->id,
+                            'question_id' =>
+                            $at_n_questtion[$rand],
                             'quizz_id' => $quiz->id,
                             // 'option_config'
                             'max_score' =>  $qt->score,
                         ];
-                        // echo "<pre>";
-                        // print_r($q->toArray());
-                        // dd($q_q);
                         $model = QuizQuestion::create($q_q);
                         $option_config = $model->option_config;
-                        // print_r($item);
-                        $o = Option::where('question_id', $q[$rand]->id)->get();
+
+                        $o = Option::where('question_id', $at_n_questtion[$rand])->get();
+
                         foreach ($o as $k => $op) {
                             $option_config[$k] = $op->id;
                         }
+
+
                         $model->option_config  =  $option_config;
 
                         $model->save();
@@ -294,14 +330,17 @@ class QuizController extends Controller
     {
 
         if ($request->has('correction_upload')) {
-            $attempt = Attempt::find($request->id);
+            $attempt = Attempt::find($request->attempt_id);
+            if (!$attempt) {
+                return back();
+            }
             if ($attempt->correction_upload != null) {
                 Storage::delete($attempt->correction_upload);
             }
             $file = $request->correction_upload;
             $ext = $request->correction_upload->extension();
-            $file->move(public_path('lms'), $request->id .  $ext);
-            $url = 'public/lms/' . $request->id .  $ext;
+            $file->move(public_path('lms'), $attempt->id . '.' .  $ext);
+            $url = 'public/lms/' . $attempt->id . '.' .  $ext;
             $attempt->correction_upload = $url;
             $attempt->save();
         }
